server: Some work on robot code.

+++ src/server/player.h
@@ -711,7 +711,7 @@ static inline bool Is_cannon_id(int id)
 void Pick_startpos(player_t *pl);
 void Go_home(player_t *pl);
 void Compute_sensor_range(player_t *pl);
-void Player_add_tank(player_t *pl, long tank_fuel);
+void Player_add_tank(player_t *pl, double tank_fuel);
 void Player_remove_tank(player_t *pl, int which_tank);
 void Player_hit_armor(player_t *pl);
 void Player_used_kill(player_t *pl);
